+++ test/browser/separate_metadata_later.cpp
@@ -12,13 +12,13 @@ EMSCRIPTEN_KEEPALIVE extern "C" void finish() {
 int main() {
   EM_ASM({
     setTimeout(function() {
-      // hijack run dep logic to see when the metadata is loaded ok.
-      var real = Module["removeRunDependency"];
-      Module["removeRunDependency"] = function(id) {
-        real(id);
-        if (id === "more.js.metadata") {
+      // hijack run blocker logic to see when the metadata is loaded ok.
+      var real = Module["addRunBlocker"];
+      Module["addRunBlocker"] = function(promise) {
+        real(promise);
+        promise.then(function() {
           Module["_finish"]();
-        }
+        });
       };
       function loadChildScript(name, then) {
         var js = document.createElement("script");